+++ src/interfaces/IPriceAllocation.sol
@@ -3,5 +3,5 @@ pragma solidity 0.8.20;
 
 interface IPriceAllocation {
     function getVestingType() external view returns (uint256);
-    function updatePrice(uint128 _newPrice) external;
+    function updatePrice(uint256 _newPrice) external;
 }
